restructure items page
