+++ frontend/src/components/item-shop/constants/icon-base.constant.ts
@@ -1,0 +1,2 @@
+/** Data Dragon's CDN root; icon URLs are `{base}/{version}/img/item/{file}`. */
+export const ICON_BASE = "https://ddragon.leagueoflegends.com/cdn";
