+++ frontend/src/components/item-shop/types/i-ddragon-raw-item-file.type.ts
@@ -1,0 +1,11 @@
+import type { IDDragonRawItem } from "./i-ddragon-raw-item.type";
+
+/**
+ * The bundled Data Dragon `item.json` asset. `tree` is the client's tag
+ * taxonomy, which the page does not read — the stat rail is its own list.
+ */
+export interface IDDragonRawItemFile {
+	version: string;
+	data: Record<string, IDDragonRawItem>;
+	tree: { header: string; tags: string[] }[];
+}
