+++ frontend/src/components/item-shop/types/i-tier-group.type.ts
@@ -1,0 +1,8 @@
+import type { IShopItem } from "./i-shop-item.type";
+import type { ItemTier } from "./item-tier.type";
+
+/** A tier heading and the items shelved under it, already sorted. */
+export interface ITierGroup {
+	tier: ItemTier;
+	items: IShopItem[];
+}
