+++ frontend/src/components/item-shop/types/item-sort.type.ts
@@ -1,0 +1,2 @@
+/** Order within a tier group. The only axis the shop sorts on. */
+export type ItemSort = "gold-asc" | "gold-desc";
