+++ frontend/vite.config.ts
@@ -6,6 +6,11 @@ import { defineConfig } from "vite";
 
 export default defineConfig({
 	plugins: [react(), tailwindcss()],
+	build: {
+		// Emit item-tag icons as separate files instead of base64-inlining them
+		// into the entry chunk.
+		assetsInlineLimit: 0,
+	},
 	resolve: {
 		alias: {
 			"@": fileURLToPath(new URL("./src", import.meta.url)),
